test-data-generator/requirements.txt
click==8.2.1
dataclasses-json==0.6.7
python-dotenv==1.1.0
requests==2.32.0
